requirements.txt
# CityBus Bot — Production Requirements

# ─── Core ───
python-telegram-bot[job-queue]>=20.0
fastapi
uvicorn[standard]
pydantic
python-dotenv

# ─── Data ───
gtfs-realtime-bindings
rapidfuzz

# ─── Database ───
motor                           # MongoDB async driver
pymongo                         # MongoDB sync (for scripts)
redis[hiredis]                  # Redis async with C parser

# ─── HTTP ───
requests

# ─── MCP ───
mcp[cli]


# ─── Testing ───
pytest
httpx
